Refactored, added tests
