+++ src/sasctl/utils/pyml2ds/connectors/ensembles/__init__.py
@@ -1,4 +1,3 @@
 from .xgb import XgbParser
 from .lgb import LightgbmParser
 from .pmml import PmmlParser
-from .forest import ForestParser
