+++ src/sasctl/utils/pyml2ds/connectors/ensembles/pmml.py
@@ -3,7 +3,7 @@ try:
 except ImportError:
     objectify = None
 
-from pyml2ds.basic.tree import TreeParser
+from pyml2ds.basic import TreeParser
 from .core import EnsembleParser
