+++ src/sasctl/utils/pyml2ds/connectors/ensembles/xgb.py
@@ -1,5 +1,5 @@
 import json
-from pyml2ds.basic.tree import TreeParser
+from pyml2ds.basic import TreeParser
 from .core import EnsembleParser
